+++ src/App.css
@@ -26,7 +26,3 @@
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
-
-.privacy ul, .terms ul{
-  padding-left:2em;
-}
